feat: consolidate situation and conversation API handling, removing deprecated files and introducing new mutation hooks
