+++ src/components/layout/dashboard-layout.tsx
@@ -1,12 +1,12 @@
 import { Button } from "@/components/ui/button";
-import DashboardSidebar from "@/domains/dashboard/components/Dashboard-Sidebar";
+import DashboardSidebar from "@/domains/dashboard-common/components/Dashboard-Sidebar";
 import { useLogout } from "@/lib/auth/hooks";
 import { ReactNode, Suspense } from "react";
 import { Link, useLocation } from "react-router-dom";
-import { useSidebarStore } from "@/domains/dashboard/store/sidebar-store";
+import { useSidebarStore } from "@/domains/dashboard-common/store/sidebar-store";
 import { Menu } from "lucide-react";
-import { useGetUsersMy } from "@/domains/dashboard/api/get-users-my";
-import { SkeletonUserProfile } from "@/domains/dashboard/components/Contents-skeleton-loading";
+import { useGetUsersMy } from "@/domains/dashboard-common/api/get-users-my";
+import { SkeletonUserProfile } from "@/domains/dashboard-common/components/Contents-skeleton-loading";
 
 interface DashboardLayoutProps {
   children: ReactNode;
