+++ src/domains/dashboard-common/components/Contents-skeleton-loading.tsx
