+++ src/domains/dashboard-conversations/api/get-all-conversations.ts
