+++ src/domains/dashboard-situations/api/delete-situations.ts
