+++ src/domains/dashboard-situations/components/modal/EditSituationModal.tsx
@@ -10,7 +10,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
-import { TAllList } from "@/domains/situation-builder/reducer/types";
+import { TAllList } from "@/domains/situation-create/reducer/types";
 import { useUpdateSituation } from "../../api/update-situation";
 import { toast } from "sonner";
 import { useQueryClient } from "@tanstack/react-query";
