+++ src/domains/dashboard-situations/hooks/use-situation-grid.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useGetInfiniteSituations } from "../api/get-situations";
-import { TAllList } from "@/domains/situation-builder/reducer/types";
+import { TAllList } from "@/domains/situation-create/reducer/types";
 import { TSituationMode } from "@/types/api";
 
 type TSituation = TAllList & { id: number; createdAt: Date };
