+++ src/domains/situation-create/api/recommend-situations.ts
