+++ src/domains/situation-create/reducer/types.ts
